app/app/requirements.txt
tornado==4.5.2
google-cloud==0.30.0
